refactor: Move hades stuff to core.net.hades.
Also, PlayerRelationsModel should not be responsible for updating HadersUserModel
