+++ common/src/main/java/com/wynntils/core/net/hades/model/HadesModel.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.sockets.model;
+package com.wynntils.core.net.hades.model;
 
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.Managers;
@@ -20,9 +20,9 @@ import com.wynntils.hades.protocol.packets.client.HCPacketUpdateStatus;
 import com.wynntils.hades.protocol.packets.client.HCPacketUpdateWorld;
 import com.wynntils.mc.event.ClientTickEvent;
 import com.wynntils.mc.utils.McUtils;
-import com.wynntils.sockets.HadesClientHandler;
-import com.wynntils.sockets.events.SocketEvent;
-import com.wynntils.sockets.objects.PlayerStatus;
+import com.wynntils.core.net.hades.HadesClientHandler;
+import com.wynntils.core.net.hades.event.HadesEvent;
+import com.wynntils.core.net.hades.objects.PlayerStatus;
 import com.wynntils.wynn.event.AthenaLoginEvent;
 import com.wynntils.wynn.event.CharacterUpdateEvent;
 import com.wynntils.wynn.event.RelationsUpdateEvent;
@@ -69,7 +69,7 @@ public final class HadesModel extends Model {
     @SubscribeEvent
     public void onAthenaLoginEvent(AthenaLoginEvent event) {
         // Try to log in to Hades, if we're not already connected
-        if (!isSocketOpen()) {
+        if (!isConnected()) {
             tryCreateConnection();
         }
     }
@@ -97,25 +97,25 @@ public final class HadesModel extends Model {
     }
 
     @SubscribeEvent
-    public void onAuth(SocketEvent.Authenticated event) {
+    public void onAuth(HadesEvent.Authenticated event) {
         pingScheduler = Executors.newSingleThreadScheduledExecutor();
         pingScheduler.scheduleAtFixedRate(this::sendPing, 0, MS_PER_PING, TimeUnit.MILLISECONDS);
     }
 
     @SubscribeEvent
-    public void onDisconnect(SocketEvent.Disconnected event) {
+    public void onDisconnect(HadesEvent.Disconnected event) {
         pingScheduler.shutdown();
     }
 
     private void sendPing() {
-        if (!isSocketOpen()) return;
+        if (!isConnected()) return;
 
         hadesConnection.sendPacketAndFlush(new HCPacketPing(System.currentTimeMillis()));
     }
 
     @SubscribeEvent
     public void onFriendListUpdate(RelationsUpdateEvent.FriendList event) {
-        if (!HadesFeature.INSTANCE.shareWithFriends || !isSocketOpen()) return;
+        if (!HadesFeature.INSTANCE.shareWithFriends || !isConnected()) return;
 
         hadesConnection.sendPacket(new HCPacketSocialUpdate(
                 event.getChangedPlayers().stream().toList(),
@@ -125,7 +125,7 @@ public final class HadesModel extends Model {
 
     @SubscribeEvent
     public void onPartyListUpdate(RelationsUpdateEvent.PartyList event) {
-        if (!HadesFeature.INSTANCE.shareWithParty || !isSocketOpen()) return;
+        if (!HadesFeature.INSTANCE.shareWithParty || !isConnected()) return;
 
         hadesConnection.sendPacket(new HCPacketSocialUpdate(
                 event.getChangedPlayers().stream().toList(),
@@ -138,7 +138,7 @@ public final class HadesModel extends Model {
         if (event.getNewState() == WorldStateManager.State.WORLD && firstWorldJoin) {
             firstWorldJoin = false;
 
-            if (!isSocketOpen()) {
+            if (!isConnected()) {
                 // FIXME: Use the proper reload command here, once they are reworked
                 MutableComponent failed = Component.literal("Welps! Trying to connect to Hades failed.")
                         .withStyle(ChatFormatting.GREEN);
@@ -163,7 +163,7 @@ public final class HadesModel extends Model {
 
     @SubscribeEvent
     public void onTick(ClientTickEvent.End event) {
-        if (!isSocketOpen()) return;
+        if (!isConnected()) return;
         if (!Managers.WorldState.onWorld() || McUtils.player().hasEffect(MobEffects.NIGHT_VISION)) return;
         if (!HadesFeature.INSTANCE.shareWithParty
                 && !HadesFeature.INSTANCE.shareWithGuild
@@ -214,7 +214,7 @@ public final class HadesModel extends Model {
     }
 
     public void tryResendWorldData() {
-        if (!isSocketOpen()) return;
+        if (!isConnected()) return;
 
         hadesConnection.sendPacket(new HCPacketUpdateWorld(
                 Managers.WorldState.getCurrentWorldName(),
@@ -222,12 +222,12 @@ public final class HadesModel extends Model {
     }
 
     public void resetSocialType(SocialType socialType) {
-        if (!isSocketOpen()) return;
+        if (!isConnected()) return;
 
         hadesConnection.sendPacketAndFlush(new HCPacketSocialUpdate(List.of(), PacketAction.RESET, socialType));
     }
 
-    private boolean isSocketOpen() {
+    private boolean isConnected() {
         return hadesConnection != null && hadesConnection.isOpen();
     }
 }
