+++ common/src/main/java/com/wynntils/core/net/hades/objects/HadesUser.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.sockets.objects;
+package com.wynntils.core.net.hades.objects;
 
 import com.wynntils.hades.protocol.packets.server.HSPacketUpdateMutual;
 import com.wynntils.mc.objects.CommonColors;
