+++ common/src/main/java/com/wynntils/wynn/model/map/poi/PlayerMiniMapPoi.java
@@ -8,7 +8,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.features.user.map.MinimapFeature;
 import com.wynntils.gui.render.RenderUtils;
 import com.wynntils.mc.utils.PlayerInfoUtils;
-import com.wynntils.sockets.objects.HadesUser;
+import com.wynntils.core.net.hades.objects.HadesUser;
 import net.minecraft.resources.ResourceLocation;
 
 public class PlayerMiniMapPoi extends PlayerPoiBase {
